web_app_rest_api/requirements.txt
numpy==1.19.5
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
torch==1.7.0
torchvision==0.8.1
albumentations==0.5.2
pillow==8.3.2